Refactor engine
