+++ Engine/include/Bored.hpp
@@ -1,4 +1,8 @@
 #pragma once
 
-#include "../src/DataAccess/FileManager/FileManager.h"
-#include "../src/GameLoop.h"
+#include "../src/DataAccess/FileManager/Manager.h"
+#include "../src/GameLoop.hpp"
+
+#if FILE_SYSTEM_API == STDFS_API
+#include "../src/DataAccess/FileManager/STDFM.h"
+#endif
