--- Engine/src/DataAccess/FileManager/FileManager.cpp
@@ -1,7 +1,0 @@
-#include "FileManager.h"
-#include "../../Adapter/STDFS.h"
-#include <memory>
-
-namespace Bored {
-namespace FileSystem {}  // namespace FileSystem
-}  // namespace Bored
