+++ Engine/src/DataAccess/FileManager/Manager.cpp
@@ -1,0 +1,7 @@
+#include "Manager.h"
+#include <memory>
+
+#if FILE_SYSTEM_API == STDFS_API
+#include "STDFM.cpp"
+#endif
+
