+++ Engine/src/GameLoop.cpp
@@ -1,35 +1,35 @@
-#include "GameLoop.h"
+#include "GameLoop.hpp"
 #include <chrono>
 
 namespace Bored {
-GameLoop::GameLoop() {}
+    GameLoop::GameLoop() {}
 
-GameLoop::~GameLoop() {}
+    GameLoop::~GameLoop() {}
 
-void GameLoop::AddModule(std::shared_ptr<Module> module) {
-  mods.push_back(module);
-}
-void GameLoop::Run() {
-  for (auto mod : mods) {
-    mod->OnSetup();
-  }
-
-  bool isRunning = true;
-  double dt = 0;
-  std::chrono::time_point<std::chrono::steady_clock> start, end;
-
-  while (isRunning) {
-    start = std::chrono::steady_clock::now();
-    for (auto mod : mods) {
-      isRunning &= mod->OnUpdate(dt);
+    void GameLoop::AddModule(std::shared_ptr<Module> module) {
+        mods.push_back(module);
     }
-    end = std::chrono::steady_clock::now();
-    const std::chrono::duration<double, std::milli> diff = end - start;
-    dt = diff.count();
-  }
 
-  for (auto mod : mods) {
-    mod->OnShutdown();
-  }
-}
+    void GameLoop::Run() {
+        for (auto mod : mods) {
+            mod->OnSetup();
+        }
+
+        bool isRunning = true;
+        double dt = 0;
+        std::chrono::steady_clock::time_point start, end;
+
+        while (isRunning) {
+            start = std::chrono::steady_clock::now();
+            for (auto mod : mods) {
+                end = std::chrono::steady_clock::now();
+                dt = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
+                isRunning &= mod->OnUpdate(dt);
+            }
+        }
+
+        for (auto mod : mods) {
+            mod->OnShutdown();
+        }
+    }
 }  // namespace Bored
